requirements.txt
Flask==2.0.1
Pillow==8.3.2
requests==2.26.0